# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: M-PESA Business Loans API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/customers:
    get:
      tags: [Admin]
      operationId: listCustomers
      summary: List all customers
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [active, suspended]
      responses:
        '200':
          description: List of customers.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CustomerSummary'
  /admin/customers/{id}:
    get:
      tags: [Admin]
      operationId: getCustomerDetail
      summary: Get a customer's full detail (profile, documents, loans)
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: Customer detail.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerDetail'
        '404':
          description: Not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /admin/customers/{id}/status:
    patch:
      tags: [Admin]
      operationId: updateCustomerStatus
      summary: Suspend or activate a customer account
      parameters:
        - $ref: '#/components/parameters/IdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CustomerStatusUpdate'
      responses:
        '200':
          description: Customer updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerSummary'
        '404':
          description: Not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /admin/loan-applications:
    get:
      tags: [Admin]
      operationId: listAllLoanApplications
      summary: List all loan applications (optionally filtered by status)
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [pending, approved, rejected, hold]
      responses:
        '200':
          description: List of loan applications with customer info.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/LoanApplicationWithCustomer'
  /admin/loan-applications/{id}/decision:
    patch:
      tags: [Admin]
      operationId: decideLoanApplication
      summary: Approve, reject, or hold a loan application
      parameters:
        - $ref: '#/components/parameters/IdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoanApplicationDecision'
      responses:
        '200':
          description: Loan application decided.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoanApplicationWithCustomer'
        '404':
          description: Not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
...
  /admin/settings:
    get:
      tags: [Settings]
      operationId: listSystemSettings
      summary: List system settings
      responses:
        '200':
          description: List of system settings.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SystemSetting'
    put:
      tags: [Settings]
      operationId: updateSystemSetting
      summary: Create or update a system setting
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SystemSettingInput'
      responses:
        '200':
          description: Setting saved.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SystemSetting'
components:
  parameters:
...
    IdParam:
      name: id
      in: path
      required: true
      schema:
        type: string
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string

    # ----- Storage -----
...
    AccountStatus:
      type: string
      enum: [active, suspended]

    # ----- Profile -----
    CustomerProfile:
      type: object
      required:
        - id
        - userId
        - phone
        - phoneVerified
        - dateOfBirth
        - address
        - city
        - nationalIdNumber
        - idFrontUrl
        - idBackUrl
        - selfieUrl
        - profileComplete
        - createdAt
        - updatedAt
      properties:
        id: { type: string }
        userId: { type: string }
        phone: { type: [string, 'null'] }
        phoneVerified: { type: boolean }
        dateOfBirth: { type: [string, 'null'] }
        address: { type: [string, 'null'] }
        city: { type: [string, 'null'] }
        nationalIdNumber: { type: [string, 'null'] }
        idFrontUrl: { type: [string, 'null'] }
        idBackUrl: { type: [string, 'null'] }
        selfieUrl: { type: [string, 'null'] }
        profileComplete: { type: boolean }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
...
    DocumentType:
      type: string
      enum: [id_front, id_back, selfie, supporting]
    Document:
      type: object
      required: [id, customerId, type, fileName, fileUrl, uploadedAt]
      properties:
        id: { type: string }
        customerId: { type: string }
        type:
          $ref: '#/components/schemas/DocumentType'
        fileName: { type: string }
        fileUrl: { type: string }
        uploadedAt: { type: string, format: date-time }
...
    LoanApplicationStatus:
      type: string
      enum: [pending, approved, rejected, hold]
    LoanApplication:
      type: object
      required:
        - id
        - customerId
        - amount
        - purpose
        - loanType
        - termMonths
        - status
        - reviewedBy
        - reviewedAt
        - reviewNotes
        - createdAt
        - updatedAt
      properties:
        id: { type: string }
        customerId: { type: string }
        amount: { type: string }
        purpose: { type: string }
        loanType: { type: string }
        termMonths: { type: integer }
        status:
          $ref: '#/components/schemas/LoanApplicationStatus'
        reviewedBy: { type: [string, 'null'] }
        reviewedAt: { type: [string, 'null'], format: date-time }
        reviewNotes: { type: [string, 'null'] }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
...
    LoanApplicationWithCustomer:
      allOf:
        - $ref: '#/components/schemas/LoanApplication'
        - type: object
          required: [customerName, customerEmail]
          properties:
            customerName: { type: [string, 'null'] }
            customerEmail: { type: [string, 'null'] }
    LoanApplicationDecision:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [approved, rejected, hold]
        reviewNotes:
          type: string
    LoanStatus:
      type: string
      enum: [active, repaid, overdue, defaulted]
    Loan:
      type: object
      required:
        - id
        - applicationId
        - customerId
        - principal
        - interestRate
        - termMonths
        - status
        - disbursedAt
        - dueDate
        - createdAt
        - updatedAt
      properties:
        id: { type: string }
        applicationId: { type: string }
        customerId: { type: string }
        principal: { type: string }
        interestRate: { type: string }
        termMonths: { type: integer }
        status:
          $ref: '#/components/schemas/LoanStatus'
        disbursedAt: { type: string, format: date-time }
        dueDate: { type: string }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
...
    CustomerSummary:
      type: object
      required:
        - id
        - email
        - firstName
        - lastName
        - accountStatus
        - phone
        - phoneVerified
        - profileComplete
        - createdAt
      properties:
        id: { type: string }
        email: { type: [string, 'null'] }
        firstName: { type: [string, 'null'] }
        lastName: { type: [string, 'null'] }
        accountStatus:
          $ref: '#/components/schemas/AccountStatus'
        phone: { type: [string, 'null'] }
        phoneVerified: { type: boolean }
        profileComplete: { type: boolean }
        createdAt: { type: string, format: date-time }
    CustomerDetail:
      allOf:
        - $ref: '#/components/schemas/CustomerSummary'
        - type: object
          required: [profile, documents, loanApplications, loans]
          properties:
            profile:
              oneOf:
                - $ref: '#/components/schemas/CustomerProfile'
                - type: 'null'
            documents:
              type: array
              items:
                $ref: '#/components/schemas/Document'
            loanApplications:
              type: array
              items:
                $ref: '#/components/schemas/LoanApplication'
            loans:
              type: array
              items:
                $ref: '#/components/schemas/Loan'
    CustomerStatusUpdate:
      type: object
      required: [accountStatus]
      properties:
        accountStatus:
          $ref: '#/components/schemas/AccountStatus'
...
    SystemSetting:
      type: object
      required: [key, value, updatedAt]
      properties:
        key: { type: string }
        value: { type: string }
        updatedAt: { type: string, format: date-time }
    SystemSettingInput:
      type: object
      required: [key, value]
      properties:
        key: { type: string, minLength: 1 }
        value: { type: string }