# lib/api-zod/src/generated/types/customerSummary.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * M-PESA Business Loans API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AccountStatus } from './accountStatus';

export interface CustomerSummary {
  id: string;
  /** @nullable */
  email: string | null;
  /** @nullable */
  firstName: string | null;
  /** @nullable */
  lastName: string | null;
  accountStatus: AccountStatus;
  /** @nullable */
  phone: string | null;
  phoneVerified: boolean;
  profileComplete: boolean;
  createdAt: Date;
}