# lib/api-zod/src/generated/types/errorEnvelope.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * M-PESA Business Loans API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ErrorEnvelope {
  error: string;
}