# lib/api-zod/src/generated/types/loanApplication.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * M-PESA Business Loans API specification
 * OpenAPI spec version: 0.1.0
 */
import type { LoanApplicationStatus } from './loanApplicationStatus';

export interface LoanApplication {
  id: string;
  customerId: string;
  amount: string;
  purpose: string;
  loanType: string;
  termMonths: number;
  status: LoanApplicationStatus;
  /** @nullable */
  reviewedBy: string | null;
  /** @nullable */
  reviewedAt: Date | null;
  /** @nullable */
  reviewNotes: string | null;
  createdAt: Date;
  updatedAt: Date;
}