# lib/api-zod/src/generated/types/loanApplicationStatus.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * M-PESA Business Loans API specification
 * OpenAPI spec version: 0.1.0
 */

export type LoanApplicationStatus = typeof LoanApplicationStatus[keyof typeof LoanApplicationStatus];


export const LoanApplicationStatus = {
  pending: 'pending',
  approved: 'approved',
  rejected: 'rejected',
  hold: 'hold',
} as const;