# lib/api-zod/src/generated/types/loanStatus.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * M-PESA Business Loans API specification
 * OpenAPI spec version: 0.1.0
 */

export type LoanStatus = typeof LoanStatus[keyof typeof LoanStatus];


export const LoanStatus = {
  active: 'active',
  repaid: 'repaid',
  overdue: 'overdue',
  defaulted: 'defaulted',
} as const;